Split schema definitions under 'tree' into multiple files
