+++ src/api/tree/__tests__/utils.test.ts
@@ -8,7 +8,7 @@ import {
   sortFoldersFirst,
 } from "../utils";
 
-import { NodeSchemaType } from "../schema";
+import { NodeSchemaType } from "../schema/node";
 
 const itemTemplate: NodeSchemaType = {
   createdAt: "2023-03-13T13:39:05.832664Z",
