+++ src/api/tree/mutate/createWorkflow.ts
@@ -1,4 +1,4 @@
-import { WorkflowCreatedSchema } from "../schema";
+import { WorkflowCreatedSchema } from "../schema/node";
 import { apiFactory } from "~/api/apiFactory";
 import { forceLeadingSlash } from "../utils";
 import { useApiKey } from "~/util/store/apiKey";
