+++ src/api/tree/mutate/deleteVariable.ts
@@ -1,7 +1,7 @@
 import {
   WorkflowVariableDeletedSchema,
   WorkflowVariableSchemaType,
-} from "../schema";
+} from "../schema/workflowVariable";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 import { apiFactory } from "~/api/apiFactory";
