+++ src/api/tree/mutate/setVariable.ts
@@ -2,7 +2,7 @@ import {
   WorkflowVariableCreatedSchema,
   WorkflowVariableCreatedSchemaType,
   WorkflowVariableFormSchemaType,
-} from "../schema";
+} from "../schema/workflowVariable";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 import { apiFactory } from "~/api/apiFactory";
