+++ src/api/tree/query/variableContent.ts
@@ -1,6 +1,6 @@
 import type { QueryFunctionContext } from "@tanstack/react-query";
 import type { ResponseParser } from "../../apiFactory";
-import { WorkflowVariableContentSchema } from "../schema";
+import { WorkflowVariableContentSchema } from "../schema/workflowVariable";
 import { apiFactory } from "../../apiFactory";
 import { forceLeadingSlash } from "../utils";
 import { treeKeys } from "..";
