+++ src/api/tree/utils.ts
@@ -1,4 +1,4 @@
-import { NodeSchemaType } from "./schema";
+import { NodeSchemaType } from "./schema/node";
 
 export const forceLeadingSlash = (path?: string) => {
   if (!path) {
