+++ src/pages/namespace/Explorer/Tree/NewWorkflow/index.tsx
@@ -21,7 +21,7 @@ import FormErrors from "~/componentsNext/FormErrors";
 import Input from "~/design/Input";
 import { Textarea } from "~/design/TextArea";
 import { addYamlFileExtension } from "./utils";
-import { fileNameSchema } from "~/api/tree/schema";
+import { fileNameSchema } from "~/api/tree/schema/node";
 import { pages } from "~/util/router/pages";
 import { useCreateWorkflow } from "~/api/tree/mutate/createWorkflow";
 import { useNamespace } from "~/util/store/namespace";
